+++ ext_localconf.php
@@ -21,6 +21,10 @@ call_user_func(function () {
         'run'
     );
 
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass']['hugo'] = 'SourceBroker\\Hugo\\DataHandling\\DataHandler';
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processCmdmapClass']['hugo'] = 'SourceBroker\\Hugo\\DataHandling\\DataHandler';
+
+
 //    Plugin indexer - TODO
 //    $signalSlotDispatcher->connect(
 //        \SourceBroker\Hugo\Traversing\PageTraverser::class,