+++ Classes/Service/HugoExportService.php
@@ -26,6 +26,7 @@ namespace SourceBroker\Hugo\Service;
 
 use SourceBroker\Hugo\Configuration\Configurator;
 use SourceBroker\Hugo\Traversing\TreeTraverser;
+use Symfony\Component\Process\Exception\ProcessFailedException;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
 use TYPO3\CMS\Core\Database\Query\Restriction\HiddenRestriction;
@@ -34,6 +35,7 @@ use TYPO3\CMS\Core\Locking\LockFactory;
 use TYPO3\CMS\Core\Locking\LockingStrategyInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
+use Symfony\Component\Process\Process;
 
 /**
  * Class HugoExportService
@@ -45,6 +47,7 @@ class HugoExportService
      * @return bool
      * @throws \TYPO3\CMS\Core\Locking\Exception\LockAcquireException
      * @throws \TYPO3\CMS\Core\Locking\Exception\LockCreateException
+     * @throws \Exception
      */
     public function exportTree(): bool
     {
@@ -78,6 +81,22 @@ class HugoExportService
                 $treeTraverser->setWriter($writer);
                 $treeTraverser->start($siteRoot['uid'], []);
             }
+        }
+        // TODO: make Process working for finding hugo path and then executing hugo itself.
+        // TODO: For now its working only for finding "which hugo". Then error on hugo build.
+        // TODO: Funny is that if Process will get output of "which hugo" then exec() is not working.
+//        $process = new Process('which hugo');
+//        $process->mustRun();
+//        if (!$process->isSuccessful()) {
+//            throw new ProcessFailedException($process);
+//        }
+//        $hugoBinaryPath = $process->getOutput();
+        $hugoBinaryPath = 'hugo';
+        if (!empty($hugoBinaryPath)) {
+            $command = $hugoBinaryPath . ' -s' . PATH_site . 'typo3conf/ext/local/Resources/Private/ClickdummyExport/site/ -d' . PATH_site . 'typo3conf/ext/local/Resources/Private/ClickdummyExport/dist';
+            exec($command);
+        } else {
+            throw new \Exception('Can not find hugo binnary');
         }
         if ($locked) {
             $locker->release();