+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 1) [TASK] Rename PageTraverser to TreeTraverser
 2) [TASK] Rename metaData to frontMatter in Document class.
 3) [TASK] Add support to disable tree export - its possible now to now export for some site root trees.
+4) [TASK] Rename values of hugo menu identifiers in TYPO3.
 
 0.0.4
 ~~~~~